remove circ dep rpc and testrpc
